// src/Types/NavItems.tsx
export interface NavItemsTypes {
  id: number;
  name: string;
  path: string;
}

export interface ContactTypes {
  name: string;
  email: string;
  description: string;
}

export interface WeddingCard {
  image: string;
  title: string;
  description: string;
  date: string;
  postDate: string;
  createdAt: number;
  _id: number;
}